[Fix]削除ボタンを微調整
